requirements.txt
amqp==5.0.6
APScheduler==3.6.3
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
cachetools==4.2.2
captcha==0.3
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.1.7
django-archive==0.2.0
django-ckeditor==6.0.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-haystack==3.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-telegrambot==1.0.1
djangorestframework==3.12.4
idna==2.10
kombu==5.1.0
lxml==4.6.3
Markdown==3.3.4
mysqlclient==2.0.3
Pillow==8.1.1
prompt-toolkit==3.0.18
pysolr==3.9.0
python-dotenv==0.17.1
python-telegram-bot==13.6
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
